- Refactor of some names;  - Remove some empty lines;  - Correct API endpoint telephone bill, which was only accepting POST requests.
